requirements.txt
flask==3.0.0
matplotlib==3.8.0
protobuf==4.24.4
python-dotenv==1.0.0
pytorch_lightning==2.1.0
scikit-learn==1.3.1
seaborn==0.13.0
sentencepiece==0.1.99
torch==2.1.0
tqdm==4.66.1
transformers==4.34.0
